\name{check.par}
\alias{check.par}
\title{
Checks major characteristics of Biograph object
}
\description{
Assesses consistency between parameter values and the Biograph object (data set) loaded}
\usage{
check.par(Bdata)
}
\arguments{
  \item{Bdata}{Biograph object}
}
\details{The function checks the presence of the date format and the correct specification of a set of parameters.   }
\author{
Frans Willekens
}
