\name{pos.charstr}
\alias{pos.charstr}
\title{
Position of a character string in a string variable
}
\description{
Determines the position of a character string in a string variable
}
\usage{
pos.charstr(string, charstr)
}
\arguments{
  \item{string}{
The character string
}
  \item{charstr}{
The character variable
}
}
\value{
First occurrence}

\author{
Frans Willekens
}
