\name{stringf}
\alias{stringf}
\title{Converts character string to vector}
\description{Converts character string to vector}
\usage{
stringf(string)
}
\arguments{
  \item{string}{character string}
}
\details{Converts string}
\value{The function returns a vector with elements the characters of a character string. }
\author{Frans Willekens}
\examples{
z <- stringf("test")
}

