\name{locpath}
\alias{locpath}
\title{Location of path in Biograph object}
\description{Determines location of path in Biograph object}
\usage{locpath(survey)}
\arguments{
  \item{survey}{input data}
}
\author{Frans Willekens}
\note{The locpath function is used in Extract and other functions.}
\examples{
data(GLHS)
loc <- locpath(GLHS)
}
